SPT-1969: Renamed UserIdentityFunction part 1 (new resources)
Duplicate UserIdentityFunction to PostPhase2UserIdentityHandler, which
now follows our preferred naming strategy. The API will be switched to
point to the new function in a future PR.
