+++ src/handlers/post-phase2-user-identity-handler/stored-identity-jwt.ts
