+++ src/identity-reuse/stored-identity-validator.ts
@@ -1,4 +1,4 @@
-import { StoredIdentityJWT } from "../handlers/user-identity/stored-identity-jwt";
+import { StoredIdentityJWT } from "../handlers/post-phase2-user-identity-handler/stored-identity-jwt";
 import { getJwtSignature } from "../commons/jwt-utilities";
 import logger from "../commons/logger";
