+++ src/tests/contract-tests/consumer/sis-consumer.pact.test.ts
@@ -1,6 +1,6 @@
 import path from "node:path";
 import { PactV4, SpecificationVersion, MatchersV3 } from "@pact-foundation/pact";
-import { UserIdentityErrorResponse } from "../../../handlers/user-identity/user-identity-error-response";
+import { UserIdentityErrorResponse } from "../../../handlers/post-phase2-user-identity-handler/post-phase2-user-identity-error-response";
 import { describe, beforeEach, it, expect, beforeAll } from "vitest";
 
 const { like } = MatchersV3;
